Implemented the Design of the Dashboard and Slight Changes to Register System
Updated the dashboard side menu, so it now includes a leader board button, as well as the rest of the buttons now are fully functional (map and rewards don't work as the map and rewards page doesn't exist yet).

Implemented 3 panels for the dashboard which will be used to track the users progress in a simple environment.

Also updated the register system, so once registered it automatically logs you in and takes you to the dashboard.